pyproject.toml
[project]
name = "codeforge"
version = "0.1.0"
description = "AI Coding Agent Orchestration Platform"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "codeforge", from = "workers" }]

[tool.poetry.dependencies]
python = "^3.12"
nats-py = "^2.9"
httpx = "^0.28"
pydantic = "^2.10"
structlog = "^25.0"
pyyaml = "^6.0"
tree-sitter = "^0.25"
tree-sitter-language-pack = "^0.13"
bm25s = "^0.3"
numpy = "^2.0"
psycopg = "^3.2"
mcp = "^1.0"
deepeval = "^2.0"
agentneo = { version = "^0.1", optional = true }
scikit-learn = "^1.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0"
pytest = "^9.0"
pytest-asyncio = "^1.0"

# -- Ruff -----------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 120
src = ["workers"]

[tool.ruff.lint]
select = [
    "F",     # pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific
    "S",     # bandit security (hardcoded passwords, shell injection, insecure crypto)
    "C4",    # unnecessary comprehensions
    "C90",   # mccabe cyclomatic complexity
    "PERF",  # performance anti-patterns
    "PIE",   # misc anti-patterns (unnecessary pass, reimplemented builtins)
    "RET",   # return issues (unnecessary else/assignment before return)
    "FURB",  # modernization hints (Path.read_text, in operator, etc.)
    "LOG",   # logging best practices
    "T20",   # print statement detection
    "PT",    # pytest style consistency
]
ignore = [
    "E501",   # line length (handled by formatter)
    "N802",   # function name should be lowercase (stdlib overrides like do_GET)
    "RET504", # unnecessary assignment before return (aids debuggability)
]

[tool.ruff.lint.isort]
known-first-party = ["codeforge"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.per-file-ignores]
"workers/tests/**/*.py" = ["S101", "PT011", "S108"]
"workers/codeforge/test_*.py" = ["S101", "PT011", "S108"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# -- Pytest ---------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["workers/tests"]
asyncio_mode = "auto"
